genirq: Broaden visibility of irq_desc's nr_actions
Commit cab303be91dc ("genirq: Add sanity checks for PM options on shared
interrupt lines") amended struct irq_desc with a nr_actions field to
facilitate direct access to the number of actions sharing an interrupt
without having to traverse the action chain.  The field's motivation was
a sanity check for interrupt handling on system sleep.  Consequently the
commit constrained the field's visibility to CONFIG_PM_SLEEP=y.

However the field is useful beyond these sanity checks:  An upcoming
commit leverages it to reduce interrupt sharing on Thunderbolt
controllers by establishing a PCIe tunnel on the downstream port which
shares its interrupt with the least number of other devices.  To support
such use cases, broaden the field's visibility to CONFIG_PM_SLEEP=n.

Signed-off-by: Lukas Wunner <lukas@wunner.de>