thunderbolt: Power down controller on Macs with Thunderbolt 3
This is an incomplete commit to add runtime PM support for 2016+ Macs
with Alpine Ridge.  It needs to be completed by someone with such a
machine because there's quite a bit of fiddling and testing involved.

First, read the documentation at the top of pm_apple.c to understand
how Apple's ACPI methods for Thunderbolt power management evolved.

On Thunderbolt 3 Macs, the TRPE method still exists, hence the pm->set
pointer will be set to it and powering the controller on and off should
work as on Thunderbolt 1 and 2.

However the XRIL method no longer exists.  It was necessary on Cactus
Ridge and newer to wait until the controller has powered down before
enabling the wake GPE.  Maybe that's no longer necessary on Alpine Ridge
and in that case, tb_pm_apple_init() needs to be amended to not search
for the method if (tb->root_switch->generation >= 3) and likewise
upstream_runtime_suspend() needs to be amended to not poll until the
controller has powered down.  (Please move the for loop to a separate
function and call it if (tb->root_switch->generation < 3).)

On Thunderbolt 1 and 2 Macs the wake GPE's number can be determined by
evaluating the XRIN object.  That object no longer exists on Macs with
Thunderbolt 3.  However there is a TGPE method which notifies the
root port above the Thunderbolt controller.  Maybe that is used to
signal wakeup on hotplug.  If that theory is correct, it is necessary
to search for the root port's handle in tb_pm_apple_init().  Basically,
call pci_upstream_bridge(tb->upstream) and call ACPI_COMPANION() on the
result, but check for NULL pointers after each of the two operations.

Then call acpi_install_notify_handler(root_port_handle, ACPI_ALL_NOTIFY,
nhi_wake_notify, nhi_dev) and declare a function nhi_wake_notify() which
returns void but is otherwise identical to nhi_wake().  I'm not sure if
ACPI_ALL_NOTIFY is correct or if it should be ACPI_SYSTEM_NOTIFY or
ACPI_DEVICE_NOTIFY.  Determine this through experimentation.

Obviously, constrain the call to acpi_install_notify_handler() to
(tb->root_switch->generation >= 3) and constrain the existing call to
acpi_install_gpe_handler() to (tb->root_switch->generation < 3).  And
be sure to remove the notify handler in tb_pm_apple_fini().

On Macs with Thunderbolt 1 and 2 it's necessary to only enable the wake
GPE when the controller is powered down and disable it otherwise.
I recall that it else fired constantly.  It's unclear to me how the
situation is on Thunderbolt 3:  There's no way to mask a notification,
the only way would be to remove the handler.  Is it necessary to remove
the handler when the controller is powered up or can it remain installed
all the time?  Probably the latter.

There are also CMPE and AMPE methods which send notifications, the
latter to the NHI and XHCI.  If listening for a notification on the
root port doesn't result in wakeup on hotplug, try installing a
notification on the NHI's handle.

With Thunderbolt 3 there is an XHCI device attached to the controller's
PCIe switch, in addition to the NHI device.  Be sure to test if wakeup
on hotplug works if:

(a) Thunderbolt device is plugged in while controller is powered down
(b) USB device is plugged in while controller is powered down
(c) Thunderbolt device is plugged in while NHI device is runtime
    suspended but controller is otherwise powered up because a USB
    device is already present
(d) USB device is plugged in while XHCI device is runtime suspended but
    controller is otherwise powered up because a Thunderbolt device is
    already present

Obviously you need both a Thunderbolt device and a USB device to test
all possible combinations.

It's unclear to me how support for the GPE-based and the notification-
based wakeup mechanism can be integrated into pm_apple.c in an elegant
way.  We can worry about that once someone has gotten it to work.

Signed-off-by: Lukas Wunner <lukas@wunner.de>