net: ks8851: Avoid device name collision with smsc95xx
The two smsc95xx USB Ethernet interfaces on the RevPi Connect used to be
named "eth0" and "eth1".  However now that the ks8851 on the PiBridge is
exposed as a third Ethernet interface which is probed in parallel to the
USB bus, the USB Ethernet interfaces are named "eth0" and "eth2".

Customers may have deployed network configurations which depend on the
"eth0" and "eth1" names for the USB interfaces, so enforce assignment of
those names by using a different name prefix for ks8851 interfaces.

Signed-off-by: Lukas Wunner <lukas@wunner.de>