serial: pl011: Determine FIFO emptiness only through BUSY bit
pl011_tx_empty() determines emptiness of the TX FIFO by checking whether
either of the BUSY or TX FIFO Full bits is set in the Flag Register.

However TX FIFO Full implies BUSY, so checking the latter is sufficient.
In fact, when pl011_console_write() polls for emptiness of the TX FIFO,
it only checks BUSY.

Drop the check for TX FIFO Full and amend pl011_console_write() to call
pl011_tx_empty() instead of duplicating its content.

The check for TX FIFO Full was introduced with historic commit
157c0342e591 ("[SERIAL] Add ARM PL011 uart primecell support.") when
amba-pl011.c was first created based on the pl010 driver.  The commit
message does not provide a reason for checking TX FIFO Full, so assume
that it was introduced accidentally:
https://git.kernel.org/tglx/history/c/157c0342e591

Signed-off-by: Lukas Wunner <lukas@wunner.de>
Cc: Russell King <rmk+kernel@armlinux.org.uk>