rtc: systohc: Auto-adjust set_offset_nsec
Commit 0f295b0650c9 ("rtc: Allow rtc drivers to specify the tv_nsec
value for ntp") allowed drivers to set a set_offset_nsec instead of
the default of 0.5 s historically used for x86-style CMOS clocks.
But none of the drivers makes use of that ability.

Measure the time required to write to the RTC and use that as the
set_offset_nsec value.  Set the initial value to 0 and rearrange the
control flow in rtc_set_ntp_time() such that the measured value is
used upon the next synchronization, not just the ones after that.

Reported-by: Markus Rudolf <markus.rudolf@mpic.de>
Signed-off-by: Lukas Wunner <lukas@wunner.de>
Cc: Prarit Bhargava <prarit@redhat.com>
Cc: Jason Gunthorpe <jgg@ziepe.ca>