rtc: hctosys: Poll hw clock for end of dead second
RTCs usually do not store fractions of a second.  They are thus akin to
a "dead second" watch which marks the second with discreet jumps between
them.

When reading the hw clock on boot to set the system time, an offset of
500 msec is currently added, thus reducing the maximum deviation from
1 sec to 500 msec.

Improve by polling the clock in 100 msec intervals for up to 10 times,
thereby further reducing the maximum deviation to 100 msec.  To avoid
delaying boot time by up to 1 second, perform the readout upon rtc
probing (which may run asynchronously) rather than in an initcall.
This has the additional benefit of reducing the delay between driver
registration and setting the system time, as well as allowing readout
for hw clocks whose driver is not built-in but provided as a module.

Reported-by: Markus Rudolf <markus.rudolf@mpic.de>
Signed-off-by: Lukas Wunner <lukas@wunner.de>
Cc: Prarit Bhargava <prarit@redhat.com>
Cc: Jason Gunthorpe <jgg@ziepe.ca>