usbnet: smsc95xx: Clean up unnecessary BUG_ON() upon register access
[commit 03b3df43ce1f64b2ec6ef8cbcf2006f3938643b2 upstream]

smsc95xx_read_reg() and smsc95xx_write_reg() call BUG_ON() if the
struct usbnet pointer passed in is NULL.

The functions have just been amended to dereference the pointer on
entry.  So the kernel now oopses if the pointer is NULL, eliminating
the need for an explicit BUG_ON().

Signed-off-by: Lukas Wunner <lukas@wunner.de>
Signed-off-by: David S. Miller <davem@davemloft.net>