PCI: pciehp: Clear incorrect Power Controller Present bit
pciehp makes the (reasonable) assumption that the device in a slot is
only accessible if Slot Power is on.

Contrary to that assumption, MengXiang reports a system where the slot
device was enumerated on boot even though Slot Power is off.  Removing
the device via sysfs is impossible:  When trying to power down the slot,
__pciehp_disable_slot() bails out with an "Already disabled" message.

The slot's claim to possess a Power Controller is bogus, so clear the
Power Controller Present bit in the cached copy of the Slot Capabilities
register.  Removal via sysfs is thereby fixed.

Log an error message to alert silicon validation engineers that the
hotplug controller's firmware needs fixing.

Reported-by: MengXiang Wan <mengxiang.wan@intel.com>
Signed-off-by: Lukas Wunner <lukas@wunner.de>