lib/asn1_encoder: Add a function to encode many byte integer values.
An example is the encoding of ECC signatures used by the ECDSA
signature verification code.  A user is the new SPDM support where
raw signatures are returned by the responder.  These can then
be encoded so that we can pass them to signature_verify()
An alternative would be to teach the ecdsa code to handle "raw"
as well as X9.62 formatted signatures.

Signed-off-by: Jonathan Cameron <Jonathan.Cameron@huawei.com>