PCI: Simplify checks in pci_parent_bus_reset()
pci_parent_bus_reset() performs a Secondary Bus Reset on the parent
bridge in order to reset a specific child.

Because siblings would be affected as well, pci_parent_bus_reset() bails
out if any are present.  It checks for presence of siblings by iterating
over the bus's devices list.  That can be simplified by checking if the
child to be reset is the sole entry in the bus list.

Another opportunity for simplification is the pci_is_root_bus(dev->bus)
check, which is subsumed by the !dev->bus->self check and thus redundant.
(For SR-IOV, pci_is_root_bus() and !dev->bus->self are not equivalent,
but the latter is true and prevents a Secondary Bus Reset even though
the former is false.)

Signed-off-by: Lukas Wunner <lukas@wunner.de>