crypto: ecdsa: improve code safety
When testing on Fedora38 (x86-HW), I ran into a hang
during the boot process. Which seemed to be caused by a NULL
dereference of the encoding of the signature `req->enc`.

This patch adds a NULL check, that now allows the boot to proceed but I
suppose it fails the ecdsa_verify. Also switch to `strncmp()` to be
'safer'.

Squish into commit: 09d3a2c

Signed-off-by: Wilfred Mallawa <wilfred.mallawa@wdc.com>