thunderbolt: Mark PCIe Adapters on root switch as non-removable
Since commit 617654aae50e ("PCI / ACPI: Identify untrusted PCI devices"),
devices below Root Ports which are declared "ExternalFacingPort" in ACPI
are considered untrusted and removable.

For Platform Integrated Host Routers (where hot-plugged devices do
appear below Root Ports), this works fine.  However Esther reports that
on certain laptops, the "ExternalFacingPort" property is set on Root
Ports leading to a Standalone Host Router.  So that Host Router is
considered removable und untrusted, even though it is soldered onto the
mainboard.

The user-visible issue is that the built-in Standalone Host Router's
PCIe Upstream and Downstream Ports are incorrectly marked untrusted and
removable.

Esther specifically calls out Lenovo X1 Carbon Gen 7/8 laptops, but
reports that products by HP and other manufacturers are likewise
affected.

It seems safe to assume that PCIe Adapters on the root switch are
generally non-removable and external-facing.

So mark them as such.

The only exception would be if a Standalone Host Router is attached
externally to the system through something like CardBus, which is
highly unlikely.

Because PCIe Adapters on the root switch are only clearly identifiable
by the Thunderbolt driver, the present fix only works if and when the
Thunderbolt driver is loaded.

For good measure, mark devices below PCIe Adapters on the root switch as
removable and untrusted in case the "ExternalFacingPort" property was
neglected to be set in ACPI.

Reported-by: Esther Shimanovich <eshimanovich@chromium.org>
Closes: https://lore.kernel.org/all/20231221-thunderbolt-pci-patch-4-v4-1-2e136e57c9bc@chromium.org/
Signed-off-by: Lukas Wunner <lukas@wunner.de>