thunderbolt: Do not bind to Host Interface exposed by Device Router
A Host Interface should only be exposed by a Host Router, never by a
Device Router.  Yet devices may erroneously or maliciously expose one.

Avoid binding to it by looking at the upstream port of its switch:

On a Host Router, the upstream port is of type TB_TYPE_NHI, whereas
on a Device Router it is of type TB_TYPE_PORT.

Signed-off-by: Lukas Wunner <lukas@wunner.de>
Cc: stable@vger.kernel.org