thunderbolt: Obtain PCIe Device/Function number via Router Operation
The PCIe Device/Function number of a PCIe Adapter (which allows
associating the Adapter with a PCI device) is normally obtainable from
DROM.

But on USB4 Routers that's only the case if the DROM is TBT3-compatible.
Moreover on Platform Integrated Host Routers the DROM is optional
(USB4 DROM Specification r1.0 sec 1).

Instead, the information is obtainable with the "Get PCIe Downstream
Entry Mapping" Router Operation (USB4 v2.0 sec 8.3.1.3.1).

Add support for it.

PCIe is optional with USB4, so constrain the Router Operation to
CONFIG_PCI=y.

The Completion Data returned by the Router Operation is somewhat
underspec'ed:  Experimentation has shown that the Non-FPB Routing ID
contains a Bus number in the upper 8 bits and a Device/Function number
in the lower 8 bits.  However the Bus number is only non-zero as long as
a tunnel has been established.  Fortunately the algorithm to associate
PCIe Adapters with PCI devices does not need the Bus number.

The spec uses the term "Native PCIe Link" without further explanation:
On Standalone Host Routers, the PCIe Downstream Ports leading to the NHI
or XHCI constitute Native PCIe Links and the Router Operation returns
(useless) entries for them.  Platform Integrated Host Routers do not
possess Native PCIe Links.

Signed-off-by: Lukas Wunner <lukas@wunner.de>